Back/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==2.3.0
pydantic-settings==2.0.3
pandas==2.0.1
openpyxl==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
google-generativeai==0.3.1